Move code for rendering timeseries as CSV to separate file
Also move corresponding unit tests to separate file and clean up 
duplicated code for rendering as CSV.
