+++ lib/forecast_utils.js
@@ -15,6 +15,7 @@ const log = bunyan.createLogger({
 })
 
 const cd2u = require('./cosmo-d2_utils')
+const tsCsv = require('./timeseries_as_csv')
 
 const secInHour = 3600
 const milliseconds = 1000
@@ -161,12 +162,6 @@ async function readMosmixTimeseriesData (
   return resultArray
 }
 
-function renderForecastCSVCosmoMosmix (voiConfigs, timeseriesDataArray) {
-  const csvLabels = gu.getCSVLabels(voiConfigs, timeseriesDataArray)
-  const csvStrings = gu.getStringsFromStationDataPayload(timeseriesDataArray)
-  return csvLabels + csvStrings.join('\n')
-}
-
 function dropTimeseriesDataNotOfInteresWithParameter (
   voiConfigs,
   timeseriesDataCollection,
@@ -195,7 +190,7 @@ function getConfigFromModel (model, COSMO_DATA_BASE_PATH, MOSMIX_DATA_BASE_PATH)
       MODEL_DATA_PATH: MOSMIX_DATA_BASE_PATH,
       unitsConverter: gu.convertUnitsForMosmix,
       jsonRenderer: renderJSONCosmoMosmix,
-      csvRenderer: renderForecastCSVCosmoMosmix
+      csvRenderer: tsCsv.renderTimeseriesAsCSV
     }
   } else if (model === 'cosmo-d2') {
     return {
@@ -205,7 +200,7 @@ function getConfigFromModel (model, COSMO_DATA_BASE_PATH, MOSMIX_DATA_BASE_PATH)
       MODEL_DATA_PATH: COSMO_DATA_BASE_PATH,
       unitsConverter: gu.convertUnitsForCosmo,
       jsonRenderer: renderJSONCosmoMosmix,
-      csvRenderer: renderForecastCSVCosmoMosmix
+      csvRenderer: tsCsv.renderTimeseriesAsCSV
     }
   } else {
     return {}
@@ -214,7 +209,6 @@ function getConfigFromModel (model, COSMO_DATA_BASE_PATH, MOSMIX_DATA_BASE_PATH)
 
 exports.renderJSONCosmoMosmix = renderJSONCosmoMosmix
 exports.readMosmixTimeseriesData = readMosmixTimeseriesData
-exports.renderForecastCSVCosmoMosmix = renderForecastCSVCosmoMosmix
 exports.dropTimeseriesDataNotOfInteresWithParameter = dropTimeseriesDataNotOfInteresWithParameter
 exports.checkIfValidModelRun = checkIfValidModelRun
 exports.getConfigFromModel = getConfigFromModel
