+++ lib/measured_values_utils.js
@@ -85,90 +85,9 @@ function renderMeasuredValuesAsJSON (voiConfigs, timeseriesDataArray, vois, sid)
   return result
 }
 
-// Functions for building CSV-representation ---------------------------------------- //
-
-function getCSVLabels (voiConfigs) {
-  let csvLabels = 'timestamp / ms,'
-  _.forEach(voiConfigs, function (voiConfig) {
-    const key = voiConfig.target.key
-    const unit = voiConfig.target.unit
-    csvLabels += `${key} / ${unit},`
-  })
-  csvLabels = _.trimEnd(csvLabels, ',') // trailing comma would indicate another column
-  csvLabels += '\n'
-  return csvLabels
-}
-
-function getHeadElementsFromTimeseries (timeseriesDataArray) {
-  // returns from each array in timeseriesDataArray first element.
-  const result = []
-  _.forEach(timeseriesDataArray, function (item) {
-    result.push(_.first(item))
-  })
-  return result
-}
-
-function getTailElementsFromTimeseries (timeseriesDataArray) {
-  const result = []
-  _.forEach(timeseriesDataArray, function (item) {
-    result.push(_.tail(item))
-  })
-  return result
-}
-
-function getValuesAsCSVString (timeStampsAndValues) {
-  // Accepts the results from getHeadElementsFromTimeseries
-  // timeStampsAndValues [{timestamp: 1, value: 1}, {...}]
-  // item {timestamp: 1, value:1}
-  // returns csv: timestamp,value1,value2
-  const timestampValue = timeStampsAndValues[0].timestamp
-  let result = `${timestampValue}`
-  _.forEach(timeStampsAndValues, function (item) {
-    result += `,${item.value}`
-  })
-  return result
-}
-
-function getStringsFromStationDataPayload (timeseriesDataArray) {
-  // TimeseriesDataArray is an array of arrays with timeseries for each voi.
-  function helper (timeseries, result) {
-    /* Iterate columnwise
-    FROM:
-    [[voi1_timeseries_el1, voi1_timeseries_el2],
-     [voi2_timeseries_el1, voi2_timeseries_el2],
-     [voi3_timeseries_el1, voi3_timeseries_el2]]
-    TO:
-    [[voi1_timeseries_el1, voi2_timeseries_el1, ... voiN...],
-     [voi1_timeseries_el2, voi2_timeseries_el2m, ... voiN]]
-     And extract values from each elements via getValuesAsCSVString
-    */
-
-    // Termination of loop condition
-    if (_.isEmpty(getHeadElementsFromTimeseries(timeseries)[0])) {
-      return result
-    }
-
-    const stringWithValues = getHeadElementsFromTimeseries(timeseries)
-    result.push(getValuesAsCSVString(stringWithValues))
-    return helper(getTailElementsFromTimeseries(timeseries), result)
-  }
-  return helper(timeseriesDataArray, [])
-}
-
-function renderMeasuredValuesAsCSV (voiConfigs, timeseriesDataArray) {
-  const csvLabels = getCSVLabels(voiConfigs, timeseriesDataArray)
-  const csvStrings = getStringsFromStationDataPayload(timeseriesDataArray)
-  return csvLabels + csvStrings.join('\n')
-}
 
 // Export functions
-exports.renderMeasuredValuesAsCSV = renderMeasuredValuesAsCSV
 exports.renderMeasuredValuesAsJSON = renderMeasuredValuesAsJSON
 exports.dropTimeseriesDataNotOfInterest = dropTimeseriesDataNotOfInterest
 exports.dropNaN = dropNaN
 exports.convertUnits = convertUnits
-
-// Export functions for testing
-exports.getHeadElementsFromTimeseries = getHeadElementsFromTimeseries
-exports.getTailElementsFromTimeseries = getTailElementsFromTimeseries
-exports.getValuesAsCSVString = getValuesAsCSVString
